+++ galaxy_ng/app/dynaconf_hooks/cors_hooks.py
@@ -1,0 +1,17 @@
+import os
+from typing import Any, Dict
+from dynaconf import Dynaconf
+
+
+def configure_cors(settings: Dynaconf) -> Dict[str, Any]:
+    """This adds CORS Middleware, useful to access swagger UI on dev environment"""
+
+    if os.getenv("DEV_SOURCE_PATH", None) is None:
+        # Only add CORS if we are in dev mode
+        return {}
+
+    data = {}
+    if settings.get("GALAXY_ENABLE_CORS", default=False):
+        corsmiddleware = ["galaxy_ng.app.common.openapi.AllowCorsMiddleware"]
+        data["MIDDLEWARE"] = corsmiddleware + settings.get("MIDDLEWARE", [])
+    return data
