+++ galaxy_ng/app/dynaconf_hooks/keycloak_hooks.py
@@ -1,0 +1,132 @@
+import os
+from typing import Any, Dict
+from dynaconf import Dynaconf
+
+
+def configure_keycloak(settings: Dynaconf) -> Dict[str, Any]:
+    """Configure keycloak settings for galaxy.
+
+    This function returns a dictionary that will be merged to the settings.
+    """
+
+    data = {}
+
+    # Obtain values for Social Auth
+    SOCIAL_AUTH_KEYCLOAK_KEY = settings.get("SOCIAL_AUTH_KEYCLOAK_KEY", default=None)
+    SOCIAL_AUTH_KEYCLOAK_SECRET = settings.get("SOCIAL_AUTH_KEYCLOAK_SECRET", default=None)
+    SOCIAL_AUTH_KEYCLOAK_PUBLIC_KEY = settings.get("SOCIAL_AUTH_KEYCLOAK_PUBLIC_KEY", default=None)
+    KEYCLOAK_PROTOCOL = settings.get("KEYCLOAK_PROTOCOL", default=None)
+    KEYCLOAK_HOST = settings.get("KEYCLOAK_HOST", default=None)
+    KEYCLOAK_PORT = settings.get("KEYCLOAK_PORT", default=None)
+    KEYCLOAK_REALM = settings.get("KEYCLOAK_REALM", default=None)
+
+    # Add settings if Social Auth values are provided
+    if all(
+        [
+            SOCIAL_AUTH_KEYCLOAK_KEY,
+            SOCIAL_AUTH_KEYCLOAK_SECRET,
+            SOCIAL_AUTH_KEYCLOAK_PUBLIC_KEY,
+            KEYCLOAK_HOST,
+            KEYCLOAK_PORT,
+            KEYCLOAK_REALM,
+        ]
+    ):
+        data["KEYCLOAK_ADMIN_ROLE"] = settings.get("KEYCLOAK_ADMIN_ROLE", default="hubadmin")
+        data["KEYCLOAK_GROUP_TOKEN_CLAIM"] = settings.get(
+            "KEYCLOAK_GROUP_TOKEN_CLAIM", default="group"
+        )
+        data["KEYCLOAK_ROLE_TOKEN_CLAIM"] = settings.get(
+            "KEYCLOAK_GROUP_TOKEN_CLAIM", default="client_roles"
+        )
+        data["KEYCLOAK_HOST_LOOPBACK"] = settings.get("KEYCLOAK_HOST_LOOPBACK", default=None)
+        data["KEYCLOAK_URL"] = f"{KEYCLOAK_PROTOCOL}://{KEYCLOAK_HOST}:{KEYCLOAK_PORT}"
+        auth_url_str = "{keycloak}/auth/realms/{realm}/protocol/openid-connect/auth/"
+        data["SOCIAL_AUTH_KEYCLOAK_AUTHORIZATION_URL"] = auth_url_str.format(
+            keycloak=data["KEYCLOAK_URL"], realm=KEYCLOAK_REALM
+        )
+        if data["KEYCLOAK_HOST_LOOPBACK"]:
+            loopback_url = "{protocol}://{host}:{port}".format(
+                protocol=KEYCLOAK_PROTOCOL, host=data["KEYCLOAK_HOST_LOOPBACK"], port=KEYCLOAK_PORT
+            )
+            data["SOCIAL_AUTH_KEYCLOAK_AUTHORIZATION_URL"] = auth_url_str.format(
+                keycloak=loopback_url, realm=KEYCLOAK_REALM
+            )
+
+        data[
+            "SOCIAL_AUTH_KEYCLOAK_ACCESS_TOKEN_URL"
+        ] = f"{data['KEYCLOAK_URL']}/auth/realms/{KEYCLOAK_REALM}/protocol/openid-connect/token/"
+
+        data["SOCIAL_AUTH_LOGIN_REDIRECT_URL"] = settings.get(
+            "SOCIAL_AUTH_LOGIN_REDIRECT_URL", default="/ui/"
+        )
+        data["SOCIAL_AUTH_POSTGRES_JSONFIELD"] = True
+        data["SOCIAL_AUTH_JSONFIELD_CUSTOM"] = "django.contrib.postgres.fields.JSONField"
+        data["SOCIAL_AUTH_URL_NAMESPACE"] = "social"
+        data["SOCIAL_AUTH_KEYCLOAK_EXTRA_DATA"] = [
+            ("refresh_token", "refresh_token"),
+            (data["KEYCLOAK_ROLE_TOKEN_CLAIM"], data["KEYCLOAK_ROLE_TOKEN_CLAIM"]),
+        ]
+
+        data["SOCIAL_AUTH_PIPELINE"] = (
+            "social_core.pipeline.social_auth.social_details",
+            "social_core.pipeline.social_auth.social_uid",
+            "social_core.pipeline.social_auth.social_user",
+            "social_core.pipeline.user.get_username",
+            "social_core.pipeline.social_auth.associate_by_email",
+            "social_core.pipeline.user.create_user",
+            "social_core.pipeline.social_auth.associate_user",
+            "social_core.pipeline.social_auth.load_extra_data",
+            "social_core.pipeline.user.user_details",
+            "galaxy_ng.app.pipelines.user_role",
+            "galaxy_ng.app.pipelines.user_group",
+        )
+
+        # Set external authentication feature flag
+        # data["GALAXY_FEATURE_FLAGS"] = {'external_authentication': True, "dynaconf_merge": True}
+        # The next have the same effect ^
+        data["GALAXY_FEATURE_FLAGS__external_authentication"] = True
+
+        # Add to installed apps
+        data["INSTALLED_APPS"] = ["social_django", "dynaconf_merge"]
+
+        # Add to authentication backends
+        data["AUTHENTICATION_BACKENDS"] = [
+            "social_core.backends.keycloak.KeycloakOAuth2",
+            "dynaconf_merge",
+        ]
+
+        # Replace AUTH CLASSES
+        data["GALAXY_AUTHENTICATION_CLASSES"] = [
+            "rest_framework.authentication.SessionAuthentication",
+            "galaxy_ng.app.auth.token.ExpiringTokenAuthentication",
+            "galaxy_ng.app.auth.keycloak.KeycloakBasicAuth"
+        ]
+
+        # Set default to one day expiration
+        data["GALAXY_TOKEN_EXPIRATION"] = settings.get("GALAXY_TOKEN_EXPIRATION", 1440)
+
+        # Add to templates
+        # Pending dynaconf issue:
+        # https://github.com/rochacbruno/dynaconf/issues/299#issuecomment-900616706
+        # So we can do a merge of this data.
+        data["TEMPLATES"] = [
+            {
+                "BACKEND": "django.template.backends.django.DjangoTemplates",
+                "DIRS": [os.path.join(settings.BASE_DIR, "templates")],
+                "APP_DIRS": True,
+                "OPTIONS": {
+                    "context_processors": [
+                        # BEGIN: Pulp standard context processors
+                        "django.template.context_processors.debug",
+                        "django.template.context_processors.request",
+                        "django.contrib.auth.context_processors.auth",
+                        "django.contrib.messages.context_processors.messages",
+                        # END: Pulp standard context processors
+                        "social_django.context_processors.backends",
+                        "social_django.context_processors.login_redirect",
+                    ],
+                },
+            },
+        ]
+
+    return data
