+++ galaxy_ng/app/dynaconf_hooks/logging_hooks.py
@@ -1,0 +1,101 @@
+import os
+from typing import Any, Dict
+from dynaconf import Dynaconf
+
+
+def configure_logging(settings: Dynaconf) -> Dict[str, Any]:
+    data = {
+        "GALAXY_ENABLE_API_ACCESS_LOG": settings.get(
+            "GALAXY_ENABLE_API_ACCESS_LOG",
+            default=os.getenv("GALAXY_ENABLE_API_ACCESS_LOG", default=False),
+        )
+    }
+    if data["GALAXY_ENABLE_API_ACCESS_LOG"]:
+        data["INSTALLED_APPS"] = ["automated_logging", "dynaconf_merge"]
+        data["MIDDLEWARE"] = [
+            "automated_logging.middleware.AutomatedLoggingMiddleware",
+            "dynaconf_merge",
+        ]
+        data["LOGGING"] = {
+            "version": 1,
+            "disable_existing_loggers": False,
+            "formatters": {
+                "automated_logging": {"format": "%(asctime)s: %(levelname)s: %(message)s"},
+            },
+            "handlers": {
+                "automated_logging": {
+                    "level": "INFO",
+                    "class": "logging.handlers.WatchedFileHandler",
+                    "filename": "/var/log/galaxy_api_access.log",
+                    "formatter": "automated_logging",
+                },
+            },
+            "loggers": {
+                "automated_logging": {
+                    "handlers": ["automated_logging"],
+                    "level": "INFO",
+                    "propagate": False,
+                },
+            },
+            "dynaconf_merge": True,
+        }
+        data["AUTOMATED_LOGGING"] = {
+            "globals": {
+                "exclude": {
+                    "applications": [
+                        "plain:contenttypes",
+                        "plain:admin",
+                        "plain:basehttp",
+                        "glob:session*",
+                        "plain:migrations",
+                    ]
+                }
+            },
+            "model": {
+                "detailed_message": True,
+                "exclude": {"applications": [], "fields": [], "models": [], "unknown": False},
+                "loglevel": 20,
+                "mask": [],
+                "max_age": None,
+                "performance": False,
+                "snapshot": False,
+                "user_mirror": False,
+            },
+            "modules": ["request", "unspecified", "model"],
+            "request": {
+                "data": {
+                    "content_types": ["application/json"],
+                    "enabled": [],
+                    "ignore": [],
+                    "mask": [
+                        "ca_cert",
+                        "client_cert",
+                        "client_key",
+                        "email",
+                        "password",
+                        "proxy_url",
+                        "proxy_username",
+                        "proxy_password",
+                        "token",
+                        "username",
+                    ],
+                    "query": True,
+                },
+                "exclude": {
+                    "applications": [],
+                    "methods": [],
+                    "status": [],
+                    "unknown": False,
+                },
+                "ip": True,
+                "loglevel": 20,
+                "max_age": None,
+            },
+            "unspecified": {
+                "exclude": {"applications": [], "files": [], "unknown": False},
+                "loglevel": 20,
+                "max_age": None,
+            },
+        }
+
+    return data
